feat(satellite_content): 更新flutter版本到1.7.8, 重构帖子内容组件(satellite_content)以支持在帖子列表和帖子详情中使用

+++ lib/components/crop_image/crop_image.dart
@@ -294,8 +294,8 @@ class CropImage extends StatelessWidget {
 
 class ImageItem {
   final String url;
-  final int width;
-  final int height;
+  final double width;
+  final double height;
 
   ImageItem({
     this.url,
